fix length

+++ Snap.Data.Mapper.Converter/HashPreConverter.cs
@@ -8,7 +8,7 @@ public class HashPreConverter : JsonConverter<HashPre>
 {
     public override HashPre Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
-        return new(reader.GetUInt16());
+        return new(reader.GetUInt64());
     }
 
     public override void Write(Utf8JsonWriter writer, HashPre value, JsonSerializerOptions options)
